+++ composer-core/src/main/java/de/otto/rx/composer/providers/ContentProvider.java
@@ -28,4 +28,5 @@ public interface ContentProvider {
      * @return Observable emitting nothing, or a single Content item.
      */
     Observable<Content> getContent(final Position position, final Tracer tracer, final Parameters parameters);
+
 }