+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## 1.0.0.M2-SNAPSHOT
 
+### New Features:
+* [Issue #17] Extracting the body part of a HTML page using `ContentProviders.htmlBodyOf(ContentProvider)`.
+
 ## 1.0.0.M1
 
 This is the first release of composer-core.